cleaning in incidence_graph sources
