+++ cgogn/core/types/incidence_graph/incidence_graph_attribute.hpp
